Make sure inline chat math wrapped in parens is rendered
Syntax like `a ($x$) b` is generated fairly often
