+++ src/vs/workbench/contrib/chat/browser/chatContentParts/chatMarkdownContentPart.ts
@@ -11,7 +11,6 @@ import { coalesce } from '../../../../../base/common/arrays.js';
 import { findLast } from '../../../../../base/common/arraysFind.js';
 import { Codicon } from '../../../../../base/common/codicons.js';
 import { Emitter } from '../../../../../base/common/event.js';
-import { Lazy } from '../../../../../base/common/lazy.js';
 import { Disposable, DisposableStore, IDisposable, MutableDisposable } from '../../../../../base/common/lifecycle.js';
 import { autorun, IObservable } from '../../../../../base/common/observable.js';
 import { equalsIgnoreCase } from '../../../../../base/common/strings.js';
@@ -35,6 +34,7 @@ import { IHoverService } from '../../../../../platform/hover/browser/hover.js';
 import { IInstantiationService } from '../../../../../platform/instantiation/common/instantiation.js';
 import { ILabelService } from '../../../../../platform/label/common/label.js';
 import { IEditorService } from '../../../../services/editor/common/editorService.js';
+import { MarkedKatexSupport } from '../../../markdown/browser/markedKatexSupport.js';
 import { IMarkdownVulnerability } from '../../common/annotations.js';
 import { IEditSessionEntryDiff } from '../../common/chatEditingService.js';
 import { IChatProgressRenderableResponseContent } from '../../common/chatModel.js';
@@ -51,7 +51,6 @@ import '../media/chatCodeBlockPill.css';
 import { IDisposableReference, ResourcePool } from './chatCollections.js';
 import { IChatContentPart, IChatContentPartRenderContext } from './chatContentParts.js';
 import { ChatExtensionsContentPart } from './chatExtensionsContentPart.js';
-import { MarkedKatexSupport } from './markedKatexSupport.js';
 import './media/chatMarkdownPart.css';
 
 const $ = dom.$;
@@ -62,67 +61,6 @@ export interface IChatMarkdownContentPartOptions {
 
 export class ChatMarkdownContentPart extends Disposable implements IChatContentPart {
 	private static idPool = 0;
-
-	private static tempSanitizerRule = new Lazy(() => {
-		// Create a CSSStyleDeclaration object via a style sheet rule
-		const styleSheet = new CSSStyleSheet();
-		styleSheet.insertRule(`.temp{}`);
-		const rule = styleSheet.cssRules[0];
-		if (!(rule instanceof CSSStyleRule)) {
-			throw new Error('Invalid CSS rule');
-		}
-		return rule.style;
-	});
-
-	private static sanitizeStyles(styleString: string, allowedProperties: readonly string[]): string {
-		const style = this.tempSanitizerRule.value;
-		style.cssText = styleString;
-
-		const sanitizedProps = [];
-
-		for (let i = 0; i < style.length; i++) {
-			const prop = style[i];
-			if (allowedProperties.includes(prop)) {
-				const value = style.getPropertyValue(prop);
-				// Allow through lists of numbers with units or bare words like 'block'
-				// Main goal is to block things like 'url()'.
-				if (/^(([\d\.\-]+\w*\s?)+|\w+)$/.test(value)) {
-					sanitizedProps.push(`${prop}: ${value}`);
-				}
-			}
-		}
-
-		return sanitizedProps.join('; ');
-	}
-
-	private static sanitizeKatexStyles(styleString: string): string {
-		const allowedProperties = [
-			'display',
-			'position',
-			'font-family',
-			'font-style',
-			'font-weight',
-			'font-size',
-			'height',
-			'width',
-			'margin',
-			'padding',
-			'top',
-			'left',
-			'right',
-			'bottom',
-			'vertical-align',
-			'transform',
-			'border',
-			'color',
-			'white-space',
-			'text-align',
-			'line-height',
-			'float',
-			'clear',
-		];
-		return this.sanitizeStyles(styleString, allowedProperties);
-	}
 
 	public readonly codeblocksPartId = String(++ChatMarkdownContentPart.idPool);
 	public readonly domNode: HTMLElement;
@@ -183,21 +121,7 @@ export class ChatMarkdownContentPart extends Disposable implements IChatContentP
 			};
 
 			const result = this._register(renderer.render(markdown.content, {
-				sanitizerOptions: {
-					allowedTags: [
-						...dom.basicMarkupHtmlTags,
-						...dom.trustedMathMlTags,
-					],
-					customAttrSanitizer: (attrName, attrValue) => {
-						if (attrName === 'class') {
-							return true; // TODO: allows all classes for now since we don't have a list of possible katex classes
-						} else if (attrName === 'style') {
-							return ChatMarkdownContentPart.sanitizeKatexStyles(attrValue);
-						}
-
-						return false;
-					},
-				},
+				sanitizerOptions: MarkedKatexSupport.getSanitizerOptions(),
 				fillInIncompleteTokens,
 				codeBlockRendererSync: (languageId, text, raw) => {
 					const isCodeBlockComplete = !isResponseVM(context.element) || context.element.isComplete || !raw || codeblockHasClosingBackticks(raw);
@@ -412,7 +336,7 @@ export class EditorPool extends Disposable {
 	}
 }
 
-function codeblockHasClosingBackticks(str: string): boolean {
+export function codeblockHasClosingBackticks(str: string): boolean {
 	str = str.trim();
 	return !!str.match(/\n```+$/);
 }
